exception handler

+++ src/main/java/com/vlad/rest/Resttest1/services/StudentService.java
@@ -2,6 +2,8 @@ package com.vlad.rest.Resttest1.services;
 
 import com.vlad.rest.Resttest1.entities.Student;
 import com.vlad.rest.Resttest1.repos.StudentRepo;
+import com.vlad.rest.Resttest1.response.RestApiException;
+import com.vlad.rest.Resttest1.response.RestApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,11 +23,21 @@ public class StudentService {
         return studentRepo.findAll();
     }
 
-    public void add(Student student) {
+    public void add(Student student) throws RestApiException {
+        if (student.getName() == null) {
+            throw new RestApiException("Field name cannot be empty!");
+        }
+        if (student.getBdate() == null) {
+            throw new RestApiException("Field bdate cannot be empty!");
+        }
+
         studentRepo.save(student);
     }
 
-    public void delete(long id) {
+    public void delete(long id) throws RestApiException {
+        if (studentRepo.findById(id).isEmpty()) {
+            throw new RestApiException("Student with id "+ id +" is not present!");
+        }
         studentRepo.deleteById(id);
     }
 
@@ -41,5 +53,13 @@ public class StudentService {
             }
             studentRepo.save(item);
         }
+    }
+
+    public Student studentById(long id) throws RestApiException {
+        Optional<Student> row = studentRepo.findById(id);
+        if (row.isEmpty()) {
+            throw new RestApiException("Student is not present!");
+        }
+        return row.get();
     }
 }
